LR2SkinLoader: Move GROOVEGAUGE to LR2SkinCSVLoader
